+++ version/113/paymentmethod/HeidelpayGiropay.php
@@ -1,0 +1,53 @@
+<?php declare(strict_types=1);
+// @phpcs:ignoreFile PSR1.Files.SideEffects.FoundWithSymbols
+
+use UnzerSDK\Resources\PaymentTypes\BasePaymentType;
+use UnzerSDK\Resources\TransactionTypes\AbstractTransactionType;
+use UnzerSDK\Resources\TransactionTypes\Charge;
+use Plugin\s360_heidelpay_shop4\Payments\HeidelpayPaymentMethod;
+use Plugin\s360_heidelpay_shop4\Payments\Interfaces\RedirectPaymentInterface;
+use Plugin\s360_heidelpay_shop4\Payments\Traits\HasCustomer;
+use Plugin\s360_heidelpay_shop4\Payments\Traits\HasMetadata;
+
+require_once __DIR__ . '/../init.php';
+
+/**
+ * Heidelpay Giropay Payment Method.
+ *
+ * Giropay is an Internet payment System in Germany, based on online banking.
+ * Introduced in February 2006, this payment method allows customers to buy
+ * securely on the Internet using direct online transfers from their bank account.
+ *
+ * Giropay is the official online banking implementation of the German banks.
+ *
+ * @see https://docs.heidelpay.com/docs/giropay-payment
+ */
+class HeidelpayGiropay extends HeidelpayPaymentMethod implements RedirectPaymentInterface
+{
+    use HasMetadata;
+    use HasCustomer;
+
+    /**
+     * @inheritDoc
+     * @return AbstractTransactionType|Charge
+     */
+    protected function performTransaction(BasePaymentType $payment, $order): AbstractTransactionType
+    {
+        // Create / Update existing customer resource if needed
+        $customer = $this->createOrFetchHeidelpayCustomer($this->adapter, $this->plugin->getSession(), false);
+
+        if ($customer->getId()) {
+            $customer = $this->adapter->getApi()->updateCustomer($customer);
+        }
+
+        return $this->adapter->getApi()->charge(
+            round($order->fGesamtsummeKundenwaehrung, 2),
+            $order->Waehrung->cISO,
+            $payment->getId(),
+            $this->getReturnURL($order),
+            $customer,
+            $order->cBestellNr ?? null,
+            $this->createMetadata()
+        );
+    }
+}